+++ assets/js/components/tracks/dot_track.ts
@@ -6,14 +6,6 @@ import { DataTrack } from "./base_tracks/data_track";
 export class DotTrack extends DataTrack {
   startExpanded: boolean;
 
-  // FIXME: Look into color bands next
-  private _colorBands: RenderBand[] | null = null;
-
-  updateColors(colorBands: RenderBand[] | null) {
-    this._colorBands = colorBands;
-    this.render({});
-  }
-
   constructor(
     id: string,
     label: string,
@@ -64,21 +56,6 @@ export class DotTrack extends DataTrack {
     const dotsInRange = dots.filter(
       (dot) => dot.x >= xRange[0] && dot.y <= xRange[1],
     );
-
-    // FIXME: Try adding Canvas bands instead
-    // let coloredDots = [...dotsInRange];
-    // if (this.colorBands != null) {
-    //   console.log("Looping the bands", coloredDots);
-    //   for (const band of this.colorBands) {
-    //     coloredDots
-    //       .filter((dot) => pointInRange(dot.x, [band.start, band.end]))
-    //       .forEach((dot) => {
-    //         if (band.color) {
-    //           dot.color = band.color;
-    //         }
-    //       });
-    //   }
-    // }
 
     drawDotsScaled(this.ctx, dotsInRange, xScale, yScale);