+++ assets/js/components/tracks_manager.ts
@@ -8,18 +8,16 @@ import { IdeogramTrack } from "./tracks/ideogram_track";
 import { OverviewTrack } from "./tracks/overview_track";
 import { DotTrack } from "./tracks/dot_track";
 import { BandTrack } from "./tracks/band_track";
-import { ANIM_TIME, COLORS, ICONS, SIZES, STYLE } from "../constants";
+import { ANIM_TIME, COLORS, SIZES, STYLE } from "../constants";
 import {
   getAnnotationContextMenuContent,
   getGenesContextMenuContent,
   getVariantContextMenuContent,
 } from "./util/menu_content_utils";
 import { ShadowBaseElement } from "./util/shadowbaseelement";
-import { generateID, populateSelect } from "../util/utils";
+import { generateID } from "../util/utils";
 import {
   getSimpleButton,
-  getContainer,
-  getIconButton,
 } from "./util/menu_utils";
 import { DataTrack } from "./tracks/base_tracks/data_track";
 import { diff, moveElement } from "../util/collections";
@@ -32,6 +30,7 @@ import {
 } from "./tracks/base_tracks/interactive_tools";
 import { getLinearScale } from "../draw/render_utils";
 import { keyLogger } from "./util/keylogger";
+import { TrackPage } from "./side_menu/track_page";
 
 const COV_Y_RANGE: [number, number] = [-3, 3];
 const BAF_Y_RANGE: [number, number] = [0, 1];
@@ -97,10 +96,12 @@ export class TracksManager extends ShadowBaseElement {
   dataSource: RenderDataSource;
   renderAll: (settings: RenderSettings) => void;
 
-  getAnnotationSources: () => { id: string; label: string }[];
+  getAnnotationSources: GetAnnotSources;
   getAnnotationDetails: (id: string) => Promise<ApiAnnotationDetails>;
   openTrackContextMenu: (track: DataTrack) => void;
   session: GensSession;
+
+  trackPages: Record<string, TrackPage> = {};
 
   constructor() {
     super(template);
@@ -124,7 +125,7 @@ export class TracksManager extends ShadowBaseElement {
 
   // FIXME: Group the callbacks for better overview
   async initialize(
-    renderAll: (settings: RenderSettings) => void,
+    render: (settings: RenderSettings) => void,
     sampleIds: string[],
     chromSizes: Record<string, number>,
     chromClick: (chrom: string) => void,
@@ -150,7 +151,7 @@ export class TracksManager extends ShadowBaseElement {
     this.getXRange = getXRange;
     this.getChromosome = getChromosome;
     this.session = session;
-    this.renderAll = renderAll;
+    this.renderAll = render;
 
     Sortable.create(this.tracksContainer, {
       animation: ANIM_TIME.medium,
@@ -159,7 +160,7 @@ export class TracksManager extends ShadowBaseElement {
         const { oldIndex, newIndex } = evt;
         const [moved] = this.dataTracks.splice(oldIndex, 1);
         this.dataTracks.splice(newIndex, 0, moved);
-        renderAll({});
+        render({});
       },
     });
 
@@ -172,19 +173,49 @@ export class TracksManager extends ShadowBaseElement {
     };
 
     this.openTrackContextMenu = (track: DataTrack) => {
-
       const isDotTrack = track instanceof DotTrack;
 
-      const returnElements = getTrackContextMenuContent(
-        renderAll,
-        track,
-        isDotTrack,
-        getAnnotSources,
-        (id: string) => dataSource.getAnnotation(id),
-        this.moveTrack,
-      );
+      if (this.trackPages[track.id] == null) {
+        const trackPage = new TrackPage();
 
-      this.session.openContextMenu(track.label, returnElements);
+        const trackSettings = {
+          showYAxis: isDotTrack,
+          showColor: true,
+        };
+
+        trackPage.configure(track.id, trackSettings);
+        this.trackPages[track.id] = trackPage;
+      }
+
+      const trackPage = this.trackPages[track.id];
+
+      this.session.showContent(track.label, [trackPage]);
+
+      trackPage.initialize(
+        getAnnotSources,
+        (direction: "up" | "down") => this.moveTrack(track.id, direction),
+        () => {
+          track.toggleHidden(), render({});
+        },
+        () => {
+          track.toggleCollapsed(), render({});
+        },
+        () => track.getIsHidden(),
+        () => track.getIsCollapsed(),
+        isDotTrack ? () => track.getYAxis().range : null,
+        (newY: Rng) => {
+          track.updateYAxis(newY);
+          render({});
+        },
+        async (annotId: string | null) => {
+          let colorBands = [];
+          if (annotId != null) {
+            colorBands = await dataSource.getAnnotation(annotId);
+          }
+          track.setColorBands(colorBands);
+          render({});
+        },
+      );
     };
 
     const covTracks = [];
@@ -254,7 +285,7 @@ export class TracksManager extends ShadowBaseElement {
     const overviewTrackBaf = this.getOverviewTrack(
       "overview_baf",
       "Overview (baf)",
-      () => dataSource.getOverviewCovData(sampleIds[0]),
+      () => dataSource.getOverviewBafData(sampleIds[0]),
       BAF_Y_RANGE,
       chromSizes,
       chromClick,
@@ -286,50 +317,20 @@ export class TracksManager extends ShadowBaseElement {
       track.renderLoading();
     });
 
-    this.setupDrag();
+    // this.setupDrag();
+    setupDrag(
+      this.tracksContainer,
+      session,
+      getXRange,
+      getChromosome,
+      this.dragCallbacks,
+    );
 
     this.tracksContainer.addEventListener("click", () => {
       if (keyLogger.heldKeys.Control) {
         this.dragCallbacks.onZoomOut();
       }
     });
-  }
-
-  setupDrag() {
-    const onDragEnd = (pxRangeX: Rng, _pxRangeY: Rng, shiftPress: boolean) => {
-      const xRange = this.getXRange();
-      if (xRange == null) {
-        console.error("No xRange set");
-      }
-
-      const yAxisWidth = STYLE.yAxis.width;
-
-      const pixelToPos = getLinearScale(
-        [yAxisWidth, this.tracksContainer.offsetWidth],
-        xRange,
-      );
-      const posStart = pixelToPos(Math.max(yAxisWidth, pxRangeX[0]));
-      const posEnd = pixelToPos(Math.max(yAxisWidth, pxRangeX[1]));
-
-      if (shiftPress) {
-        this.dragCallbacks.onZoomIn([Math.floor(posStart), Math.floor(posEnd)]);
-      } else {
-        const id = generateID();
-        this.dragCallbacks.addHighlight({
-          id,
-          chromosome: this.getChromosome(),
-          range: [posStart, posEnd],
-          color: COLORS.transparentBlue,
-        });
-      }
-    };
-
-    initializeDragSelect(
-      this.tracksContainer,
-      onDragEnd,
-      this.dragCallbacks.removeHighlight,
-      () => this.session.getMarkerMode(),
-    );
   }
 
   getDataTracks(): DataTrack[] {
@@ -339,24 +340,32 @@ export class TracksManager extends ShadowBaseElement {
   getTrackById(trackId: string): DataTrack {
     for (let i = 0; i < this.dataTracks.length; i++) {
       if (this.dataTracks[i].id == trackId) {
-        return this.dataTracks[i];
+        const track = this.dataTracks[i];
+        return track;
       }
     }
     console.error("Did not find any track with ID", trackId);
   }
 
+  // FIXME: Hmm, messy with the tracks containers, think about how to simplify this
   moveTrack(trackId: string, direction: "up" | "down") {
     const track = this.getTrackById(trackId);
     const trackIndex = this.dataTracks.indexOf(track);
     const shift = direction == "up" ? -1 : 1;
 
-    const container = track.parentNode;
+    const trackContainer = track.parentElement;
+
+    // const singleTrackContainer = track.parentNode;
+    const trackSContainer = trackContainer.parentNode;
     if (direction == "up") {
-      container.insertBefore(track, this.dataTracks[trackIndex - 1]);
+      trackSContainer.insertBefore(
+        trackContainer,
+        this.dataTracks[trackIndex - 1].parentNode,
+      );
     } else {
-      container.insertBefore(
-        track,
-        this.dataTracks[trackIndex + 1].nextSibling,
+      trackSContainer.insertBefore(
+        trackContainer,
+        this.dataTracks[trackIndex + 1].parentNode.nextSibling,
       );
     }
 
@@ -374,7 +383,7 @@ export class TracksManager extends ShadowBaseElement {
   }
 
   updateAnnotationTracks() {
-    const sources = this.getAnnotationSources();
+    const sources = this.getAnnotationSources({ selectedOnly: true });
     // const sourcesIds = sources.map((source) => source.id);
 
     const currAnnotTracks = this.annotationTracks;
@@ -426,6 +435,10 @@ export class TracksManager extends ShadowBaseElement {
 
     this.updateAnnotationTracks();
 
+    for (const trackPage of Object.values(this.trackPages)) {
+      trackPage.render(settings);
+    }
+
     this.ideogramTrack.render(settings);
 
     for (const track of this.dataTracks) {
@@ -449,6 +462,7 @@ export class TracksManager extends ShadowBaseElement {
     }
 
     const openContextMenuId = async (id: string) => {
+      console.log("opening context menu");
       const details = await this.getAnnotationDetails(id);
       const button = getSimpleButton("Set highlight", () => {
         const id = generateID();
@@ -462,7 +476,7 @@ export class TracksManager extends ShadowBaseElement {
       const entries = getAnnotationContextMenuContent(id, details);
       const content = [button];
       content.push(...entries);
-      () => this.session.openContextMenu("Annotations", content);
+      this.session.showContent("Annotations", content);
     };
 
     const track = new BandTrack(
@@ -552,7 +566,7 @@ export class TracksManager extends ShadowBaseElement {
         const content = [button];
         content.push(...entries);
 
-        this.session.openContextMenu("Variant", content);
+        this.session.showContent("Variant", content);
       },
       this.openTrackContextMenu,
       this.session,
@@ -590,7 +604,7 @@ export class TracksManager extends ShadowBaseElement {
         const entries = getGenesContextMenuContent(id, details);
         const content = [button];
         content.push(...entries);
-        this.session.openContextMenu("Transcript", content);
+        this.session.showContent("Transcript", content);
       },
       this.openTrackContextMenu,
       this.session,
@@ -653,131 +667,47 @@ function appendDataTrack(parentContainer: HTMLDivElement, track: DataTrack) {
   parentContainer.appendChild(wrapper);
 }
 
-// FIXME: Where should this util go?
-function getTrackContextMenuContent(
-  render: (settings: RenderSettings) => void,
-  track: DataTrack,
-  isDotTrack: boolean,
-  getAnnotationSources: GetAnnotSources,
-  getBands: (id: string) => Promise<RenderBand[]>,
-  moveTrack: (id: string, direction: "up" | "down") => void,
-): HTMLDivElement[] {
-  const buttonsDiv = document.createElement("div");
-  buttonsDiv.style.display = "flex";
-  buttonsDiv.style.flexDirection = "row";
-  buttonsDiv.style.flexWrap = "nowrap";
-  buttonsDiv.style.gap = `${SIZES.s}px`;
+function setupDrag(
+  tracksContainer: HTMLElement,
+  session: GensSession,
+  getXRange: () => Rng,
+  getChromosome: () => string,
+  dragCallbacks: DragCallbacks,
+) {
+  const onDragEnd = (pxRangeX: Rng, _pxRangeY: Rng, shiftPress: boolean) => {
+    const xRange = getXRange();
+    if (xRange == null) {
+      console.error("No xRange set");
+    }
 
-  buttonsDiv.appendChild(
-    getIconButton(ICONS.up, "Up", () => moveTrack(track.id, "up")),
-  );
-  buttonsDiv.appendChild(
-    getIconButton(ICONS.down, "Down", () => moveTrack(track.id, "down")),
-  );
-  buttonsDiv.appendChild(
-    getIconButton(
-      track.getIsHidden() ? ICONS.hide : ICONS.show,
-      "Show / hide",
-      () => {
-        track.toggleHidden();
-        render({});
-      },
-    ),
-  );
-  buttonsDiv.appendChild(
-    getIconButton(
-      track.getIsCollapsed() ? ICONS.maximize : ICONS.minimize,
-      "Collapse / expand",
-      () => {
-        track.toggleCollapsed();
-        render({});
-      },
-    ),
-  );
+    const yAxisWidth = STYLE.yAxis.width;
 
-  const returnElements = [buttonsDiv];
-
-  if (isDotTrack) {
-    const axisRow = getYAxisRow(render, track);
-    returnElements.push(axisRow);
-
-    const colorSelectRow = getColorSelectRow(
-      track as DotTrack,
-      getAnnotationSources,
-      getBands,
+    const pixelToPos = getLinearScale(
+      [yAxisWidth, tracksContainer.offsetWidth],
+      xRange,
     );
-    returnElements.push(colorSelectRow);
-  }
-  return returnElements;
-}
+    const posStart = pixelToPos(Math.max(yAxisWidth, pxRangeX[0]));
+    const posEnd = pixelToPos(Math.max(yAxisWidth, pxRangeX[1]));
 
-function getYAxisRow(render: (settings: RenderSettings) => void, track: DataTrack): HTMLDivElement {
-  const axis = track.getYAxis();
-
-  const axisRow = getContainer("row");
-  axisRow.style.gap = "8px";
-  const yAxisLabel = document.createTextNode(`Y-axis range:`);
-  axisRow.appendChild(yAxisLabel);
-
-  const startNode = document.createElement("input");
-  startNode.type = "number";
-  startNode.value = axis.range[0].toString();
-  startNode.step = "0.1";
-  startNode.style.flex = "1 1 0px";
-  startNode.style.minWidth = "0";
-  axisRow.appendChild(startNode);
-
-  const endNode = document.createElement("input");
-  endNode.type = "number";
-  endNode.value = axis.range[1].toString();
-  endNode.step = "0.1";
-  endNode.style.flex = "1 1 0px";
-  endNode.style.minWidth = "0";
-
-  const onRangeChange = () => {
-    const parsedRange: Rng = [
-      parseFloat(startNode.value),
-      parseFloat(endNode.value),
-    ];
-    track.updateYAxis(parsedRange);
-    render({});
+    if (shiftPress) {
+      dragCallbacks.onZoomIn([Math.floor(posStart), Math.floor(posEnd)]);
+    } else {
+      const id = generateID();
+      dragCallbacks.addHighlight({
+        id,
+        chromosome: getChromosome(),
+        range: [posStart, posEnd],
+        color: COLORS.transparentBlue,
+      });
+    }
   };
 
-  startNode.addEventListener("change", onRangeChange);
-  endNode.addEventListener("change", onRangeChange);
-  axisRow.appendChild(endNode);
-  return axisRow;
-}
-
-function getColorSelectRow(
-  track: DotTrack,
-  getAnnotationSources: GetAnnotSources,
-  getBands: (id: string) => Promise<RenderBand[]>,
-): HTMLDivElement {
-  const colorRow = getContainer("row");
-
-  const label = document.createTextNode("Color:");
-
-  const select = document.createElement("select") as HTMLSelectElement;
-  select.style.backgroundColor = COLORS.white;
-  const sources = getAnnotationSources();
-
-  populateSelect(select, sources, true);
-
-  select.addEventListener("change", async (_e) => {
-    const currentId = select.value;
-    if (currentId != "") {
-      const bands = await getBands(currentId);
-      track.updateColors(bands);
-    } else {
-      track.updateColors(null);
-    }
-  });
-
-  colorRow.appendChild(label);
-  colorRow.appendChild(select);
-
-  return colorRow;
+  initializeDragSelect(
+    tracksContainer,
+    onDragEnd,
+    dragCallbacks.removeHighlight,
+    () => session.getMarkerMode(),
+  );
 }
 
 customElements.define("gens-tracks", TracksManager);