+++ assets/js/gens.ts
@@ -3,18 +3,19 @@ import "./components/input_controls";
 import "./components/util/popup";
 import "./components/util/shadowbaseelement";
 import "./components/util/choice_select";
-import "./components/side_menu";
-import "./components/settings_page";
+import "./components/side_menu/settings_page";
+import "./components/side_menu/side_menu";
 import "./components/header_info";
 import "./components/util/marker";
+import "./components/side_menu/track_page";
 
 import { API } from "./state/api";
 import { TracksManager } from "./components/tracks_manager";
 import { InputControls } from "./components/input_controls";
 import { getRenderDataSource } from "./state/parse_data";
 import { CHROMOSOMES } from "./constants";
-import { SideMenu } from "./components/side_menu";
-import { SettingsPage } from "./components/settings_page";
+import { SideMenu } from "./components/side_menu/side_menu";
+import { SettingsPage } from "./components/side_menu/settings_page";
 import { HeaderInfo } from "./components/header_info";
 import { GensSession } from "./state/session";
 
@@ -84,12 +85,7 @@ export async function initCanvases({
 
   const session = new GensSession(render, sideMenu);
 
-  setupShortcuts(
-    session,
-    sideMenu,
-    inputControls,
-    onChromClick,
-  );
+  setupShortcuts(session, sideMenu, inputControls, onChromClick);
 
   const annotSources = await api.getAnnotationSources();
   const defaultAnnot = annotSources
@@ -135,7 +131,7 @@ export async function initCanvases({
     sideMenu.showContent("Settings", [settingsPage]);
 
     if (!settingsPage.isInitialized) {
-      settingsPage.initialize(render);
+      settingsPage.initialize();
     }
   });
 }
@@ -181,7 +177,8 @@ async function initialize(
       render({ dataUpdated: true });
     },
     () => inputControls.zoomOut(),
-    () => settingsPage.getAnnotSources(),
+    (settings: { selectedOnly: boolean }) =>
+      settingsPage.getAnnotSources(settings),
     getVariantURL,
     async (id: string) => await api.getAnnotationDetails(id),
     async (id: string) => await api.getTranscriptDetails(id),